# openapi/platform-api.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "fieldErrors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
...
      "CreateEvaluationProfileRequest": {
        "properties": {
          "aggregation": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "passThreshold": {
            "type": "string"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/EvaluationRuleInput"
            },
            "type": "array"
          },
          "visibility": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "rules"
        ],
        "type": "object"
      },
...
      "CreateSandboxProfileRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SandboxProfileVersionInput"
          },
          {
            "properties": {
              "description": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "name": {
                "type": "string"
              },
              "ownerDepartmentId": {
                "format": "uuid",
                "type": "string"
              }
            },
            "required": [
              "name",
              "ownerDepartmentId"
            ],
            "type": "object"
          }
        ]
      },
...
      "EvaluationRuleInput": {
        "properties": {
          "configuration": {},
          "evaluatorType": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "weight": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "evaluatorType",
          "configuration"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "SandboxProfileVersionInput": {
        "properties": {
          "cpuMillis": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "diskBytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "imageDigest": {
            "type": "string"
          },
          "memoryBytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "networkPolicy": {},
          "outputLimitBytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "pidsLimit": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "runner": {
            "type": "string"
          },
          "timeoutSeconds": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "runner",
          "imageDigest",
          "cpuMillis",
          "memoryBytes",
          "pidsLimit",
          "diskBytes",
          "timeoutSeconds",
          "outputLimitBytes",
          "networkPolicy"
        ],
        "type": "object"
      },
...
      "UpdateMarkdownRequest": {
        "properties": {
          "content": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "expectedRevision": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "content",
          "expectedRevision"
        ],
        "type": "object"
      },
...
      "UpdateSandboxProfileRequest": {
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "name",
          "status",
          "version"
        ],
        "type": "object"
      },
...
    },
...
  },
  "info": {
    "description": "",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Agentx Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evaluation-profiles": {
      "get": {
        "operationId": "list_profiles",
        "responses": {},
        "tags": [
          "datasets"
        ]
      },
      "post": {
        "operationId": "create_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvaluationProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "datasets"
        ]
      }
    },
...
    "/api/v1/sandbox-profiles": {
      "get": {
        "operationId": "list_sandbox_profiles",
        "responses": {},
        "tags": [
          "sandbox_profiles"
        ]
      },
      "post": {
        "operationId": "create_sandbox_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "sandbox_profiles"
        ]
      }
    },
    "/api/v1/sandbox-profiles/{id}": {
      "delete": {
        "operationId": "delete_sandbox_profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expectedVersion",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "deletion"
        ]
      },
      "get": {
        "operationId": "get_sandbox_profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "sandbox_profiles"
        ]
      },
      "patch": {
        "operationId": "update_sandbox_profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSandboxProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "sandbox_profiles"
        ]
      }
    },
    "/api/v1/sandbox-profiles/{id}/versions": {
      "post": {
        "operationId": "create_sandbox_profile_version",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SandboxProfileVersionInput"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "sandbox_profiles"
        ]
      }
    },
...
    "/api/v1/skills/{id}/files/{entry_id}": {
      "get": {
        "operationId": "get_file",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "skills_control"
        ]
      },
      "put": {
        "operationId": "update_markdown",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMarkdownRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "skills_control"
        ]
      }
    },
...
  },
...
}